Remove unnecessary CGAL_OGL_Polyhedron and VBOPolyhedron
Separate out CVAL_OGL_VBOPolyhedron implementation into separate compilation unit

Remove excessive debug output
